Move bundle compiler infra into test-helpers
