+++ packages/@glimmer/bundle-compiler/test/basic-components-test.ts
@@ -1,0 +1,3 @@
+import { rawModule, BasicComponents, BundlingRenderDelegate } from "@glimmer/test-helpers";
+
+rawModule("[Bundle Compiler] Basic Components", BasicComponents, BundlingRenderDelegate, { componentModule: true });
