--- packages/@glimmer/bundle-compiler/test/bundle-compiler-test.ts
@@ -1,1 +1,0 @@
-export {};
