+++ packages/@glimmer/bundle-compiler/test/debugger-test.ts
@@ -1,0 +1,3 @@
+import { rawModule, DebuggerSuite, BundlingRenderDelegate } from "@glimmer/test-helpers";
+
+rawModule('[Bundle Compiler] Debugger Tests', DebuggerSuite, BundlingRenderDelegate);
