+++ packages/@glimmer/bundle-compiler/test/emberish-components-test.ts
@@ -1,0 +1,3 @@
+import { rawModule, EmberishComponentTests, BundlingRenderDelegate } from "@glimmer/test-helpers";
+
+rawModule('[Bundle Compiler] Emberish Components', EmberishComponentTests, BundlingRenderDelegate, { componentModule: true });
