+++ packages/@glimmer/bundle-compiler/test/has-block-params-helper-test.ts
@@ -1,0 +1,3 @@
+import { rawModule, HasBlockParamsHelperSuite, BundlingRenderDelegate } from "@glimmer/test-helpers";
+
+rawModule('[Bundle Compiler] Has Block Params Helpers Tests', HasBlockParamsHelperSuite, BundlingRenderDelegate, { componentModule: true });
