+++ packages/@glimmer/bundle-compiler/test/has-block-test.ts
@@ -1,0 +1,3 @@
+import { rawModule, HasBlockSuite, BundlingRenderDelegate } from "@glimmer/test-helpers";
+
+rawModule('[Bundle Compiler] Has Block', HasBlockSuite, BundlingRenderDelegate, { componentModule: true });
