+++ packages/@glimmer/bundle-compiler/test/in-element-test.ts
@@ -1,0 +1,3 @@
+import { rawModule, BundlingRenderDelegate, InElementSuite } from "@glimmer/test-helpers";
+
+rawModule('[Bundle Compiler] In-Element Tests', InElementSuite, BundlingRenderDelegate);
