+++ packages/@glimmer/bundle-compiler/test/scope-test.ts
@@ -1,0 +1,3 @@
+import { rawModule, ScopeSuite, BundlingRenderDelegate } from "@glimmer/test-helpers";
+
+rawModule('[Bundle Compiler] Scope Tests', ScopeSuite, BundlingRenderDelegate, { componentModule: true });
