+++ packages/@glimmer/bundle-compiler/test/shadowing-test.ts
@@ -1,0 +1,3 @@
+import { rawModule, ShadowingSuite, BundlingRenderDelegate } from "@glimmer/test-helpers";
+
+rawModule('[Bundle Compiler] Shadowing Tests', ShadowingSuite, BundlingRenderDelegate, { componentModule: true });
