+++ packages/@glimmer/bundle-compiler/test/with-dynamic-vars-test.ts
@@ -1,0 +1,3 @@
+import { rawModule, BundlingRenderDelegate, WithDynamicVarsSuite } from "@glimmer/test-helpers";
+
+rawModule('[Bundle Compiler] With Dynamic Vars Tests', WithDynamicVarsSuite, BundlingRenderDelegate, { componentModule: true });
