+++ packages/@glimmer/bundle-compiler/test/yield-test.ts
@@ -1,0 +1,3 @@
+import { rawModule, YieldSuite, BundlingRenderDelegate } from "@glimmer/test-helpers";
+
+rawModule('[Bundle Compiler] Yield Tests', YieldSuite, BundlingRenderDelegate, { componentModule: true });
